backend/requirements.txt
# Flask core
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-SocketIO==5.3.5

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Authentication
Authlib==1.3.0
PyJWT==2.8.0

# Environment variables
python-dotenv==1.0.0

# WebSocket support
python-socketio==5.10.0
eventlet==0.33.3

# Utilities
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
python-dateutil==2.8.2

# Location services
geopy==2.4.1
haversine==2.8.0

# Development
gunicorn==21.2.0
